requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cryptography==3.4.8
cssselect==1.1.0
cssutils==2.3.0
cytoolz==0.11.0
dbus-python==1.2.16
eth-abi==2.1.1
eth-account==0.5.5
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
idna==3.2
importlib-metadata==4.6.4
ipfshttpclient==0.7.0
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.1.0
keyrings.alt==4.1.0
loguru==0.5.3
lru-dict==1.1.7
lxml==4.6.3
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
parsimonious==0.8.1
premailer==3.10.0
protobuf==3.17.3
pyarmor==6.7.4
pycparser==2.20
pycryptodome==3.10.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyrsistent==0.18.0
python-dotenv==0.19.0
requests==2.26.0
rlp==2.0.1
SecretStorage==3.3.1
six==1.16.0
toolz==0.11.1
typing-extensions==3.10.0.0
urllib3==1.26.6
varint==1.0.2
web3==5.23.0
websockets==9.1
yagmail==0.14.256
yarl==1.6.3
zipp==3.5.0
